test(integration): split MutatingEndpointTests into per-handler files; simplify PutUserInfo body read
Two follow-ups bundled:

1. Split MutatingEndpointTests.cs (1206 lines, 62 tests covering two
   distinct endpoint surfaces) into per-source-file test classes,
   mirroring the existing read-only layout (FileEndpointTests /
   ContainerEndpointTests / FolderEndpointTests / EcosystemEndpointTests):

   - test/WopiHost.IntegrationTests/Fixtures/MutatingEndpointsFixture.cs
     (the lifted Fixture + [CollectionDefinition("MutatingEndpoints")])
   - test/WopiHost.IntegrationTests/FileMutatingEndpointTests.cs
     (PutFile / RenameFile / PutRelativeFile / PutUserInfo / DeleteFile /
      ProcessCobalt / ProcessLock + Lock state machine + cross-cutting
      lock-cycle + unknown-override fallthrough)
   - test/WopiHost.IntegrationTests/ContainerMutatingEndpointTests.cs
     (CreateChildContainer / CreateChildFile / DeleteContainer /
      RenameContainer)

   Both classes share one backend boot via [Collection("MutatingEndpoints")]
   — same pattern as ReadOnlyEndpointsFixture/Collection. The fixture's
   token-mint helpers were promoted from per-class methods to public
   members on the fixture (no duplication across the two files). All
   62 tests preserved verbatim modulo s/MintFileTokenAsync(.../_fixture.MintFileTokenAsync(.../

2. Simplify PutUserInfo body read. The previous impl allocated both a
   MemoryStream and a separate `chunk` byte[] (~2 KB total) and copied
   `chunk → MemoryStream` per ReadAsync iteration. ReadAsync writes
   directly into the final buffer at the running offset:

       var buffer = new byte[PutUserInfoMaxBytes + 1];
       while ((read = await Body.ReadAsync(buffer.AsMemory(total), ct)) > 0)
       { total += read; if (total > MAX) return BadRequest(); }
       userInfo = Encoding.UTF8.GetString(buffer, 0, total);

   One allocation instead of two, no intermediate copy, same MAX+1
   over-read detection semantics. ArrayPool not used — PutUserInfo is
   rare and 1KB is gen-0 noise.

Core.Tests 313/313, IntegrationTests 114/114 (same as before the
split — no test added or removed).
